// web/src/lib/api/models/account.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Account {

  /**
   * A URL to the JSON Schema for this object.
   */
  '$schema'?: string;

  /**
   * Creation timestamp
   */
  createTime: string;

  /**
   * Short account code
   */
  displayCode: string;

  /**
   * Optional free-text description
   */
  displayDescription: string;

  /**
   * Human-readable account name
   */
  displayName: string;

  /**
   * Account UUID
   */
  id: string;

  /**
   * Whether the account is archived
   */
  isArchived: boolean;

  /**
   * Parent account UUID; null for root accounts
   */
  parentAccountId?: string;

  /**
   * Last modification timestamp
   */
  updateTime: string;
}