// web/src/lib/api/models/account-group.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AccountGroup {

  /**
   * A URL to the JSON Schema for this object.
   */
  '$schema'?: string;

  /**
   * Creation timestamp
   */
  createTime: string;

  /**
   * Optional free-text description
   */
  displayDescription: string;

  /**
   * Human-readable account group name
   */
  displayName: string;

  /**
   * Account group UUID
   */
  id: string;

  /**
   * Last modification timestamp
   */
  updateTime: string;
}