// web/src/lib/api/models/account-group-assignment.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AccountGroupAssignment {

  /**
   * A URL to the JSON Schema for this object.
   */
  '$schema'?: string;
  accountGroupId: string;
  accountId: string;
  createTime: string;
  id: string;

  /**
   * Whether to exclude this account from the group
   */
  negate: boolean;
  updateTime: string;
}