// web/src/lib/api/models/budget.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Budget {

  /**
   * A URL to the JSON Schema for this object.
   */
  '$schema'?: string;

  /**
   * Creation timestamp
   */
  createTime: string;

  /**
   * Optional free-text description
   */
  displayDescription: string;

  /**
   * Human-readable budget name
   */
  displayName: string;

  /**
   * Budget UUID
   */
  id: string;

  /**
   * Whether the budget is closed
   */
  isClosed: boolean;

  /**
   * Budget period end date
   */
  periodEnd: string;

  /**
   * Budget period start date
   */
  periodStart: string;

  /**
   * Last modification timestamp
   */
  updateTime: string;
}