// web/src/lib/api/models/budget-revision.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BudgetRevision {

  /**
   * A URL to the JSON Schema for this object.
   */
  '$schema'?: string;

  /**
   * Parent budget UUID
   */
  budgetId: string;

  /**
   * Creation timestamp
   */
  createTime: string;

  /**
   * Revision date
   */
  date: string;

  /**
   * Optional free-text description
   */
  displayDescription: string;

  /**
   * Budget revision UUID
   */
  id: string;

  /**
   * Last modification timestamp
   */
  updateTime: string;
}