// web/src/lib/api/models/import-source.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ImportSource {

  /**
   * A URL to the JSON Schema for this object.
   */
  '$schema'?: string;

  /**
   * Creation timestamp
   */
  createTime: string;

  /**
   * Optional free-text description
   */
  displayDescription: string;

  /**
   * Human-readable import source name
   */
  displayName: string;

  /**
   * Import source UUID
   */
  id: string;

  /**
   * Period start date
   */
  periodStart: string;

  /**
   * Last modification timestamp
   */
  updateTime: string;
}