// web/src/lib/api/models/import-source-period.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ImportSourcePeriod {

  /**
   * A URL to the JSON Schema for this object.
   */
  '$schema'?: string;
  createTime: string;
  id: string;
  importSourceId: string;

  /**
   * Whether this period is closed for new imports
   */
  isClosed: boolean;
  updateTime: string;
  year: number;
}