// web/src/lib/api/models/report.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Report {

  /**
   * A URL to the JSON Schema for this object.
   */
  '$schema'?: string;

  /**
   * Creation timestamp
   */
  createTime: string;

  /**
   * Human-readable name
   */
  displayName: string;

  /**
   * Report UUID
   */
  id: string;
}